refactor: consolidate ReactState class
