+++ packages/react/src/jsx-runtime.test.tsx
@@ -2,7 +2,7 @@
 
 import { act, render, screen } from '@testing-library/react';
 import React, { Children, Component, isValidElement } from 'react';
-import { Consumer, get, State, set } from '.';
+import State, { Consumer, get, set } from '.';
 
 it('will create and provide instance', () => {
   class Control extends State {
@@ -56,7 +56,7 @@ describe('new method', () => {
     const didCreate = jest.fn();
 
     class Test extends State {
-      new() {
+      protected new() {
         didCreate();
       }
     }
